+++ changelog.md
@@ -2,10 +2,11 @@
 
 Note, only relatively significant changes to user-visible functionality will be included here. Most recent changes at the top.
 
-## 20240224
+## 20240226
 
 * Refactored noise generation functions (will break seeds).
 * Added `SamplerOverride` node.
+* `studentt` noise type replaced with `studentt_test` (the more correct version).
 
 ## 20240210